% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declaretive.R
\docType{package}
\name{bindedEval}
\alias{bindedEval}
\alias{bindedEval-package}
\alias{swap_args}
\title{Evaluation with name binding}
\usage{
swap_args(f)
}
\description{
Probides evaluation with name binding mimicking Haskell's `where` clause.
}

