% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindedEval.R
\name{let}
\alias{let}
\title{Evaluates expr with bindings. The binding in rhs are locked, so}
\usage{
let(..., .in)
}
\arguments{
\item{...}{binding}

\item{.in}{expression to be evaluated}
}
\value{
the result of evaluation of `.in` with binding
}
\examples{
velocity <- function (x1, x2, t1, t2) let(
  dx = x2 - x1,
  dt = t2 - t1,
  .in. = {
    dx / dt
  }
)
velocity(80, 90, 0, 0.5)
}

