% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindedEval.R
\name{where}
\alias{where}
\title{Add binding to function.}
\usage{
where(FUN, ...)
}
\arguments{
\item{FUN}{function}

\item{...}{binding}
}
\value{
FUN with binding
}
\examples{
velocity <- (function (x1, x2, t1, t2)) \%>\%
  where(
    dx = x2 - x1,
    dt = t2 - t1)
velocity(80, 90, 0, 0.5)
}

